Implemented folder structure
